Move definition of Claim to internal.
