+++ src/internal/query/mod.rs
@@ -1,4 +1,6 @@
+pub(crate) mod claim;
 pub(crate) mod filter;
 #[cfg(feature = "parallel")]
 pub(crate) mod par_view;
 pub(crate) mod view;
+
