+++ src/public/system/schedule/builder.rs
@@ -1,6 +1,7 @@
 use crate::{
     component::Component,
     entity,
+    internal::query::claim::Claim,
     query::{
         filter, result, view,
         view::{View, Views},
@@ -147,65 +148,6 @@ impl<'a> ParSystem<'a> for Null {
         R: Registry,
     {
         unsafe { unreachable_unchecked() }
-    }
-}
-
-pub trait Claim {
-    fn claim(mutable_claims: &mut HashSet<TypeId>, immutable_claims: &mut HashSet<TypeId>);
-}
-
-impl<C> Claim for &C
-where
-    C: Component,
-{
-    fn claim(_mutable_claims: &mut HashSet<TypeId>, immutable_claims: &mut HashSet<TypeId>) {
-        immutable_claims.insert(TypeId::of::<C>());
-    }
-}
-
-impl<C> Claim for &mut C
-where
-    C: Component,
-{
-    fn claim(mutable_claims: &mut HashSet<TypeId>, _immutable_claims: &mut HashSet<TypeId>) {
-        mutable_claims.insert(TypeId::of::<C>());
-    }
-}
-
-impl<C> Claim for Option<&C>
-where
-    C: Component,
-{
-    fn claim(_mutable_claims: &mut HashSet<TypeId>, immutable_claims: &mut HashSet<TypeId>) {
-        immutable_claims.insert(TypeId::of::<C>());
-    }
-}
-
-impl<C> Claim for Option<&mut C>
-where
-    C: Component,
-{
-    fn claim(mutable_claims: &mut HashSet<TypeId>, _immutable_claims: &mut HashSet<TypeId>) {
-        mutable_claims.insert(TypeId::of::<C>());
-    }
-}
-
-impl Claim for entity::Identifier {
-    fn claim(_mutable_claims: &mut HashSet<TypeId>, _immutable_claims: &mut HashSet<TypeId>) {}
-}
-
-impl Claim for view::Null {
-    fn claim(_mutable_claims: &mut HashSet<TypeId>, _immutable_claims: &mut HashSet<TypeId>) {}
-}
-
-impl<'a, V, W> Claim for (V, W)
-where
-    V: View<'a> + Claim,
-    W: Views<'a> + Claim,
-{
-    fn claim(mutable_claims: &mut HashSet<TypeId>, immutable_claims: &mut HashSet<TypeId>) {
-        V::claim(mutable_claims, immutable_claims);
-        W::claim(mutable_claims, immutable_claims);
     }
 }
